requirements.txt
numpy==2.1.3
pillow==11.0.0
tabulate==0.9.0
